refactor(autofix): split check-suite outcome handlers
