+++ tests/sentry/seer/autofix/pr_iteration/test_ready_for_review.py
@@ -4,19 +4,18 @@ import orjson
 
 from sentry.scm.types import CheckSuiteEvent
 from sentry.seer.agent.client_models import RepoPRState, SeerRunState
-from sentry.seer.autofix.pr_iteration.check_suites import (
-    READY_FOR_REVIEW_EXTRA,
-    CheckRunsSweep,
-    CheckSuiteAutofixRun,
-    GreenCheckSuite,
-    resolve_check_suite,
-)
+from sentry.seer.autofix.pr_iteration.check_suites import CheckRunsSweep, CheckSuiteAutofixRun
 from sentry.seer.autofix.pr_iteration.constants import REVIEW_REQUEST_FLAG
-from sentry.seer.autofix.pr_iteration.ready_for_review import mark_ready_for_review
+from sentry.seer.autofix.pr_iteration.green_check_suite import (
+    READY_FOR_REVIEW_EXTRA,
+    GreenCheckSuite,
+    _mark_ready_for_review,
+)
+from sentry.seer.autofix.pr_iteration.resolve import resolve_check_suite
 from sentry.testutils.cases import TestCase
 
-READY_FOR_REVIEW_PATH = "sentry.seer.autofix.pr_iteration.ready_for_review"
-CHECK_SUITES_PATH = "sentry.seer.autofix.pr_iteration.check_suites"
+GREEN_PATH = "sentry.seer.autofix.pr_iteration.green_check_suite"
+RESOLVE_PATH = "sentry.seer.autofix.pr_iteration.resolve"
 
 RUN_ID = 67890
 REPO_NAME = "owner/repo"
@@ -80,7 +79,7 @@ def _mark_ready(event: CheckSuiteEvent | None = None) -> None:
     ctx = resolved.confirm_green()
     if ctx is None:
         return
-    mark_ready_for_review(ctx)
+    _mark_ready_for_review(ctx)
 
 
 class MarkReadyForReviewTest(TestCase):
@@ -93,19 +92,19 @@ class MarkReadyForReviewTest(TestCase):
             organization=self.organization, seer_run_state_id=RUN_ID, user_id=self.user.id
         )
         repos_patcher = patch(
-            f"{CHECK_SUITES_PATH}.resolve_check_suite_repositories", return_value=[self.repo]
+            f"{RESOLVE_PATH}.resolve_check_suite_repositories", return_value=[self.repo]
         )
         repos_patcher.start()
         self.addCleanup(repos_patcher.stop)
         self.get_pr = MagicMock(return_value=_pull_request_result())
-        get_pr_patcher = patch(f"{CHECK_SUITES_PATH}.scm_actions.get_pull_request", self.get_pr)
+        get_pr_patcher = patch(f"{GREEN_PATH}.scm_actions.get_pull_request", self.get_pr)
         get_pr_patcher.start()
         self.addCleanup(get_pr_patcher.stop)
         # MagicMock does not satisfy runtime_checkable SCM protocols.
-        proto_patcher = patch(f"{CHECK_SUITES_PATH}.GetPullRequestProtocol", object)
+        proto_patcher = patch(f"{GREEN_PATH}.GetPullRequestProtocol", object)
         proto_patcher.start()
         self.addCleanup(proto_patcher.stop)
-        sweep_patcher = patch(f"{CHECK_SUITES_PATH}.sweep_check_runs", return_value=GREEN_SWEEP)
+        sweep_patcher = patch(f"{GREEN_PATH}.sweep_check_runs", return_value=GREEN_SWEEP)
         sweep_patcher.start()
         self.addCleanup(sweep_patcher.stop)
 
@@ -129,15 +128,15 @@ class MarkReadyForReviewTest(TestCase):
         self.seer_run.refresh_from_db()
         return (self.seer_run.extras or {}).get(READY_FOR_REVIEW_EXTRA, {}).get(REPO_NAME)
 
-    @patch(f"{READY_FOR_REVIEW_PATH}.MarkPullRequestDraftStateProtocol", object)
-    @patch(f"{READY_FOR_REVIEW_PATH}.scm_actions")
+    @patch(f"{GREEN_PATH}.MarkPullRequestDraftStateProtocol", object)
+    @patch(f"{GREEN_PATH}.scm_actions.mark_pull_request_ready_for_review")
     @patch("sentry.scm.factory.new", return_value=MagicMock())
-    @patch(f"{CHECK_SUITES_PATH}.resolve_check_suite_autofix_run")
+    @patch(f"{RESOLVE_PATH}.resolve_check_suite_autofix_run")
     def test_undrafts(
         self,
         mock_resolve: MagicMock,
         _mock_scm: MagicMock,
-        mock_actions: MagicMock,
+        mock_mark: MagicMock,
     ) -> None:
         # Live PR head matches the suite even if run_state.commit_sha is stale.
         mock_resolve.return_value = self._resolved(commit_sha="stale-run-state")
@@ -148,11 +147,11 @@ class MarkReadyForReviewTest(TestCase):
         marker = self._marker()
         assert marker is not None
         assert marker["head_sha"] == HEAD_SHA
-        mock_actions.mark_pull_request_ready_for_review.assert_called_once()
-        _scm, pr_number = mock_actions.mark_pull_request_ready_for_review.call_args[0]
+        mock_mark.assert_called_once()
+        _scm, pr_number = mock_mark.call_args[0]
         assert pr_number == str(PR_NUMBER)
 
-    @patch(f"{CHECK_SUITES_PATH}.resolve_check_suite_autofix_run")
+    @patch(f"{RESOLVE_PATH}.resolve_check_suite_autofix_run")
     def test_noop_when_flag_disabled(self, mock_resolve: MagicMock) -> None:
         mock_resolve.return_value = self._resolved()
         _mark_ready(_green_event())
@@ -160,14 +159,14 @@ class MarkReadyForReviewTest(TestCase):
         mock_resolve.assert_called_once()
         assert self._marker() is None
 
-    @patch(f"{READY_FOR_REVIEW_PATH}.scm_actions")
+    @patch(f"{GREEN_PATH}.scm_actions.mark_pull_request_ready_for_review")
     @patch("sentry.scm.factory.new", return_value=MagicMock())
-    @patch(f"{CHECK_SUITES_PATH}.resolve_check_suite_autofix_run")
+    @patch(f"{RESOLVE_PATH}.resolve_check_suite_autofix_run")
     def test_skips_stale_head(
         self,
         mock_resolve: MagicMock,
         _mock_scm: MagicMock,
-        mock_actions: MagicMock,
+        mock_mark: MagicMock,
     ) -> None:
         mock_resolve.return_value = self._resolved()
         self.get_pr.return_value = _pull_request_result(head_sha="newer")
@@ -176,21 +175,21 @@ class MarkReadyForReviewTest(TestCase):
             _mark_ready(_green_event())
 
         assert self._marker() is None
-        mock_actions.mark_pull_request_ready_for_review.assert_not_called()
+        mock_mark.assert_not_called()
 
-    @patch(f"{READY_FOR_REVIEW_PATH}.scm_actions")
+    @patch(f"{GREEN_PATH}.scm_actions.mark_pull_request_ready_for_review")
     @patch("sentry.scm.factory.new", return_value=MagicMock())
     @patch(
-        f"{CHECK_SUITES_PATH}.sweep_check_runs",
+        f"{GREEN_PATH}.sweep_check_runs",
         return_value=CheckRunsSweep(total=2, incomplete=1, failed=0),
     )
-    @patch(f"{CHECK_SUITES_PATH}.resolve_check_suite_autofix_run")
+    @patch(f"{RESOLVE_PATH}.resolve_check_suite_autofix_run")
     def test_skips_when_not_green(
         self,
         mock_resolve: MagicMock,
         _mock_sweep: MagicMock,
         _mock_scm: MagicMock,
-        mock_actions: MagicMock,
+        mock_mark: MagicMock,
     ) -> None:
         mock_resolve.return_value = self._resolved()
 
@@ -198,35 +197,35 @@ class MarkReadyForReviewTest(TestCase):
             _mark_ready(_green_event())
 
         assert self._marker() is None
-        mock_actions.mark_pull_request_ready_for_review.assert_not_called()
+        mock_mark.assert_not_called()
 
-    @patch(f"{READY_FOR_REVIEW_PATH}.MarkPullRequestDraftStateProtocol", object)
-    @patch(f"{READY_FOR_REVIEW_PATH}.scm_actions")
+    @patch(f"{GREEN_PATH}.MarkPullRequestDraftStateProtocol", object)
+    @patch(f"{GREEN_PATH}.scm_actions.mark_pull_request_ready_for_review")
     @patch("sentry.scm.factory.new", return_value=MagicMock())
-    @patch(f"{CHECK_SUITES_PATH}.resolve_check_suite_autofix_run")
+    @patch(f"{RESOLVE_PATH}.resolve_check_suite_autofix_run")
     def test_undraft_failure_leaves_marker_unset(
         self,
         mock_resolve: MagicMock,
         _mock_scm: MagicMock,
-        mock_actions: MagicMock,
+        mock_mark: MagicMock,
     ) -> None:
         mock_resolve.return_value = self._resolved()
-        mock_actions.mark_pull_request_ready_for_review.side_effect = RuntimeError("boom")
+        mock_mark.side_effect = RuntimeError("boom")
 
         with self.feature(REVIEW_REQUEST_FLAG):
             _mark_ready(_green_event())
 
         assert self._marker() is None
 
-    @patch(f"{READY_FOR_REVIEW_PATH}.MarkPullRequestDraftStateProtocol", object)
-    @patch(f"{READY_FOR_REVIEW_PATH}.scm_actions")
+    @patch(f"{GREEN_PATH}.MarkPullRequestDraftStateProtocol", object)
+    @patch(f"{GREEN_PATH}.scm_actions.mark_pull_request_ready_for_review")
     @patch("sentry.scm.factory.new", return_value=MagicMock())
-    @patch(f"{CHECK_SUITES_PATH}.resolve_check_suite_autofix_run")
+    @patch(f"{RESOLVE_PATH}.resolve_check_suite_autofix_run")
     def test_skips_when_marker_exists_for_any_head(
         self,
         mock_resolve: MagicMock,
         _mock_scm: MagicMock,
-        mock_actions: MagicMock,
+        mock_mark: MagicMock,
     ) -> None:
         # Sticky: prior undraft on an older SHA must not undraft a new tip.
         self.seer_run.update(
@@ -244,20 +243,20 @@ class MarkReadyForReviewTest(TestCase):
         with self.feature(REVIEW_REQUEST_FLAG):
             _mark_ready(_green_event())
 
-        mock_actions.mark_pull_request_ready_for_review.assert_not_called()
+        mock_mark.assert_not_called()
         marker = self._marker()
         assert marker is not None
         assert marker["head_sha"] == "older-sha"
 
-    @patch(f"{READY_FOR_REVIEW_PATH}.MarkPullRequestDraftStateProtocol", object)
-    @patch(f"{READY_FOR_REVIEW_PATH}.scm_actions")
+    @patch(f"{GREEN_PATH}.MarkPullRequestDraftStateProtocol", object)
+    @patch(f"{GREEN_PATH}.scm_actions.mark_pull_request_ready_for_review")
     @patch("sentry.scm.factory.new", return_value=MagicMock())
-    @patch(f"{CHECK_SUITES_PATH}.resolve_check_suite_autofix_run")
+    @patch(f"{RESOLVE_PATH}.resolve_check_suite_autofix_run")
     def test_skips_undraft_when_pr_not_draft(
         self,
         mock_resolve: MagicMock,
         _mock_scm: MagicMock,
-        mock_actions: MagicMock,
+        mock_mark: MagicMock,
     ) -> None:
         mock_resolve.return_value = self._resolved()
         self.get_pr.return_value = _pull_request_result(draft=False)
@@ -265,20 +264,20 @@ class MarkReadyForReviewTest(TestCase):
         with self.feature(REVIEW_REQUEST_FLAG):
             _mark_ready(_green_event())
 
-        mock_actions.mark_pull_request_ready_for_review.assert_not_called()
+        mock_mark.assert_not_called()
         marker = self._marker()
         assert marker is not None
         assert marker["head_sha"] == HEAD_SHA
 
-    @patch(f"{READY_FOR_REVIEW_PATH}.MarkPullRequestDraftStateProtocol", object)
-    @patch(f"{READY_FOR_REVIEW_PATH}.scm_actions")
+    @patch(f"{GREEN_PATH}.MarkPullRequestDraftStateProtocol", object)
+    @patch(f"{GREEN_PATH}.scm_actions.mark_pull_request_ready_for_review")
     @patch("sentry.scm.factory.new", return_value=MagicMock())
-    @patch(f"{CHECK_SUITES_PATH}.resolve_check_suite_autofix_run")
+    @patch(f"{RESOLVE_PATH}.resolve_check_suite_autofix_run")
     def test_skips_undraft_when_pr_not_open(
         self,
         mock_resolve: MagicMock,
         _mock_scm: MagicMock,
-        mock_actions: MagicMock,
+        mock_mark: MagicMock,
     ) -> None:
         mock_resolve.return_value = self._resolved()
         self.get_pr.return_value = _pull_request_result(state="closed", merged=True)
@@ -286,7 +285,7 @@ class MarkReadyForReviewTest(TestCase):
         with self.feature(REVIEW_REQUEST_FLAG):
             _mark_ready(_green_event())
 
-        mock_actions.mark_pull_request_ready_for_review.assert_not_called()
+        mock_mark.assert_not_called()
         # Sticky marker so later green suites don't keep confirming + undrafting.
         marker = self._marker()
         assert marker is not None
